+++ src/Mailer/UserMailer.php
@@ -22,6 +22,7 @@
  */
 namespace MeCms\Mailer;
 
+use Cake\Network\Exception\InternalErrorException;
 use MeCms\Mailer\Mailer;
 
 /**
@@ -30,45 +31,75 @@ use MeCms\Mailer\Mailer;
 class UserMailer extends Mailer
 {
     /**
-     * Email to activate the user account account (signup and resend activation)
-     * @param array $user User data
+     * Email to activate the user account (signup and resend activation).
+     *
+     * The `$user` object must contain the `email` and `full_name` properties
+     * @param \MeCms\Model\Entity\User $user User data
      * @return void
      * @see MeCms\Controller\Admin\UsersController::resendActivation()
      * @see MeCms\Controller\Admin\UsersController::signup()
+     * @throws InternalErrorException
      */
     public function activateAccount($user)
     {
+        //Checks that all required data is present
+        foreach (['email', 'full_name'] as $property) {
+            if (empty($user->$property)) {
+                throw new InternalErrorException(__d('me_cms', 'Missing `{0}` property from data', $property));
+            }
+        }
+
         $this->to([$user->email => $user->full_name])
-            ->set(['fullName' => $user->full_name])
             ->subject(__d('me_cms', 'Activate your account'))
-            ->template('MeCms.Users/activate_account');
+            ->template('MeCms.Users/activate_account')
+            ->set(['fullName' => $user->full_name]);
     }
 
     /**
-     * Email to change the user's password
-     * @param array $user User data
+     * Email to change the user's password.
+     *
+     * The `$user` object must contain the `email` and `full_name` properties
+     * @param \MeCms\Model\Entity\User $user User data
      * @return void
      * @see MeCms\Controller\Admin\UsersController::changePassword()
+     * @throws InternalErrorException
      */
     public function changePassword($user)
     {
+        //Checks that all required data is present
+        foreach (['email', 'full_name'] as $property) {
+            if (empty($user->$property)) {
+                throw new InternalErrorException(__d('me_cms', 'Missing `{0}` property from data', $property));
+            }
+        }
+
         $this->to([$user->email => $user->full_name])
-            ->set(['fullName' => $user->full_name])
             ->subject(__d('me_cms', 'Your password has been changed'))
-            ->template('MeCms.Users/change_password');
+            ->template('MeCms.Users/change_password')
+            ->set(['fullName' => $user->full_name]);
     }
 
     /**
-     * Email to ask a new password
-     * @param array $user User data
+     * Email to ask a new password.
+     *
+     * The `$user` object must contain the `email` and `full_name` properties
+     * @param \MeCms\Model\Entity\User $user User data
      * @return void
      * @see MeCms\Controller\UsersController::forgotPassword()
+     * @throws InternalErrorException
      */
     public function forgotPassword($user)
     {
+        //Checks that all required data is present
+        foreach (['email', 'full_name'] as $property) {
+            if (empty($user->$property)) {
+                throw new InternalErrorException(__d('me_cms', 'Missing `{0}` property from data', $property));
+            }
+        }
+
         $this->to([$user->email => $user->full_name])
-            ->set(['fullName' => $user->full_name])
             ->subject(__d('me_cms', 'Reset your password'))
-            ->template('MeCms.Users/forgot_password');
+            ->template('MeCms.Users/forgot_password')
+            ->set(['fullName' => $user->full_name]);
     }
 }