+++ tests/TestCase/Model/Validation/BannerValidatorTest.php
@@ -74,10 +74,9 @@ class BannerValidatorTest extends TestCase
      */
     public function testValidationExampleData()
     {
-        $errors = $this->Banners->newEntity($this->example)->errors();
-        $this->assertEmpty($errors);
+        $this->assertEmpty($this->Banners->newEntity($this->example)->errors());
 
-        foreach ($this->example as $key => $value) {
+        foreach (array_keys($this->example) as $key) {
             //Create a copy of the example data and removes the current value
             $copy = $this->example;
             unset($copy[$key]);
@@ -120,9 +119,6 @@ class BannerValidatorTest extends TestCase
      */
     public function testValidationForTarget()
     {
-        $this->example['target'] = 'http://example.com';
-        $this->assertEmpty($this->Banners->newEntity($this->example)->errors());
-
         $this->example['target'] = 'string';
         $this->assertEquals([
             'target' => ['url' => 'Must be a valid url'],
@@ -135,6 +131,9 @@ class BannerValidatorTest extends TestCase
 
         $this->example['target'] = 'http://example.com/' . str_repeat('a', 236);
         $this->assertEmpty($this->Banners->newEntity($this->example)->errors());
+
+        $this->example['target'] = 'http://example.com';
+        $this->assertEmpty($this->Banners->newEntity($this->example)->errors());
     }
 
     /**
@@ -143,14 +142,14 @@ class BannerValidatorTest extends TestCase
      */
     public function testValidationForThumbnail()
     {
-        foreach ([true, false] as $value) {
-            $this->example['thumbnail'] = $value;
-            $this->assertEmpty($this->Banners->newEntity($this->example)->errors());
-        }
-
         $this->example['thumbnail'] = 'string';
         $this->assertEquals([
             'thumbnail' => ['boolean' => 'You have to select a valid option'],
         ], $this->Banners->newEntity($this->example)->errors());
+
+        foreach ([true, false] as $value) {
+            $this->example['thumbnail'] = $value;
+            $this->assertEmpty($this->Banners->newEntity($this->example)->errors());
+        }
     }
 }