+++ tests/TestCase/Model/Validation/BannersPositionValidatorTest.php
@@ -75,7 +75,7 @@ class BannersPositionValidatorTest extends TestCase
     {
         $this->assertEmpty($this->BannersPositions->newEntity($this->example)->errors());
 
-        foreach ($this->example as $key => $value) {
+        foreach (array_keys($this->example) as $key) {
             //Create a copy of the example data and removes the current value
             $copy = $this->example;
             unset($copy[$key]);