+++ tests/TestCase/Model/Validation/PageValidatorTest.php
@@ -78,7 +78,7 @@ class PageValidatorTest extends TestCase
     {
         $this->assertEmpty($this->Pages->newEntity($this->example)->errors());
 
-        foreach ($this->example as $key => $value) {
+        foreach (array_keys($this->example) as $key) {
             //Create a copy of the example data and removes the current value
             $copy = $this->example;
             unset($copy[$key]);