+++ tests/TestCase/Model/Validation/PhotoValidatorTest.php
@@ -76,7 +76,7 @@ class PhotoValidatorTest extends TestCase
     {
         $this->assertEmpty($this->Photos->newEntity($this->example)->errors());
 
-        foreach ($this->example as $key => $value) {
+        foreach (array_keys($this->example) as $key) {
             //Create a copy of the example data and removes the current value
             $copy = $this->example;
             unset($copy[$key]);