+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # 2.x branch
 ## 2.15 branch
+### 2.15.3
+* fixed code for `BackupForm` class. Added `_getBackupExportInstance()` method;
+* fixed code and a little bug for `ContactForm` and `ContactFormMailer` classes;
+* fixed code for `UserMailer` class;
+* added tests for all `Form` classes;
+* added tests for all `Mailer` classes.
+
 ### 2.15.2
 * added `beforeMarshal()` method do `PostsTable` class. Removed
     `buildTagsForRequestData()`;